Move recursive subform loader to cache.
